Create TotalOrdersSold2.sql
Why is the previous query output only 9 records and the required is the TOP 10?
Regarding the “Employees Table” that consists of 10 employees only, the only ID that is not found in the previous screenshot is the EmployeeID = 10. This query clarifies the query applied to the employee who has no record for the TotalOrdersSold.